requirements.txt
click==8.0.3
Flask==2.0.2
importlib-metadata==4.10.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pigpio==1.78
pkg-resources==0.0.0
python-dotenv==0.19.2
python-mpv==0.5.2
typing-extensions==4.0.1
Werkzeug==2.0.2
youtube-dl==2021.12.17
zipp==3.7.0
